requirements.txt
# Framework Web
flask==2.3.3  # Version stable pour RPi
flask-sqlalchemy==3.0.5
flask-login==0.6.2
werkzeug==2.3.7
aiohttp==3.8.5

# Base de données
sqlalchemy==2.0.20  # Version testée sur RPi
alembic==1.12.0

# Tests
pytest==7.4.2
pytest-asyncio==0.21.1

# Gestion des médias
ffmpeg-python==0.2.0
pydub==0.25.1
mutagen==1.46.0  # Version stable pour RPi
eyed3==0.9.7
m3u8==3.4.0

# Utilitaires
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
psutil==5.9.5
tenacity==8.2.3

# Sécurité
bcrypt==4.0.1
PyJWT==2.8.0
cryptography==41.0.3

# APIs et Téléchargement
yt-dlp==2023.7.6  # Version testée sur RPi
beautifulsoup4==4.12.2
lxml==4.9.3

# Audio et Interface
PyQt6==6.5.2  # Version compatible ARM
simpleaudio==1.0.4  # Alternative légère à pyaudio

# Performance et Optimisation
numpy==1.24.3  # Version optimisée pour ARM
Pillow==10.0.0  # Pour le traitement d'images
cachetools==5.3.1  # Pour le cache en mémoire

# Networking
websockets==11.0.3
dns-python==2.4.2  # Pour la résolution DNS asynchrone
